requirements.txt
asgiref==3.3.1
Django==3.1.3
django-bootstrap-modal-forms==2.0.0
django-crispy-forms==1.10.0
pytz==2020.4
sqlparse==0.4.1
